CSS cleanups and better formatting
Switched from Monseurat to Open Sans, mostly because of the condensed versions being available.
Also killes `skeleton.css`. Was using skeleton as a base to get up and running, but I was using so little of it other than the base styles, and I modified it pretty heavily, so I decided to ditch it and just copy the styles I was using.